requirements.txt
python-dotenv==1.0.1 # For reading environment variables stored in .env file
flask>=2.0.0 # Web framework for the application
waitress>=2.1.0 # Production WSGI server
langchain>=0.2.2
langchain-community>=0.2.3
langchain-google-genai>=1.0.0 # For Google Gemini integration
langchain-openai>=0.1.8 # For embeddings (keeping for comparison)
unstructured>=0.11.8 # Document loading (compatible with Python 3.13)
# onnxruntime==1.17.1 # chromadb dependency: on Mac use `conda install onnxruntime -c conda-forge`
# For Windows users, install Microsoft Visual C++ Build Tools first
# install onnxruntime before installing `chromadb`
chromadb>=0.5.0 # Vector storage
google-generativeai>=0.3.0 # Google Gemini API
openai>=1.31.1 # For embeddings (keeping for comparison)
tiktoken>=0.7.0  # For embeddings
numpy>=1.24.0 # For mathematical operations and similarity calculations

# New dependencies for Lovable.dev prompt generator
jinja2>=3.1.0 # Template engine for prompt generation
pyyaml>=6.0 # YAML configuration file support
streamlit>=1.28.0 # Web interface
# argparse is built-in, no need to install

# install markdown dependencies with: `pip install "unstructured[md]"` after install the requirements file. Leave this line commented out.
